Remove table headers from resource to component (#114)
