+++ src/stock-items/stock-items-table.component.tsx
@@ -41,7 +41,6 @@ const StockItemsTableComponent: React.FC<StockItemsTableProps> = () => {
   const {
     isLoading,
     items,
-    tableHeaders,
     totalCount,
     currentPageSize,
     pageSizes,
@@ -50,6 +49,52 @@ const StockItemsTableComponent: React.FC<StockItemsTableProps> = () => {
     isDrug,
     setDrug,
   } = useStockItemsPages(ResourceRepresentation.Full);
+
+  const tableHeaders = useMemo(
+    () => [
+      {
+        id: 0,
+        header: t("type", "Type"),
+        key: "type",
+      },
+      {
+        id: 1,
+        header: t("genericName", "Generic Name"),
+        key: "genericName",
+      },
+      {
+        id: 2,
+        header: t("commonName", "Common Name"),
+        key: "commonName",
+      },
+      {
+        id: 3,
+        header: t("tradeName", "Trade Name"),
+        key: "tradeName",
+      },
+      {
+        id: 4,
+        header: t("dispensingUnitName", "Dispensing UoM"),
+        key: "dispensingUnitName",
+      },
+      {
+        id: 5,
+        header: t("defaultStockOperationsUoMName", "Bulk Packaging"),
+        key: "defaultStockOperationsUoMName",
+      },
+      {
+        id: 6,
+        header: t("reorderLevel", "Reorder Level"),
+        key: "reorderLevel",
+      },
+      {
+        id: 7,
+        key: "actions",
+        header: "Actions",
+      },
+    ],
+    [t]
+  );
 
   const tableRows = useMemo(() => {
     return items?.map((stockItem, index) => ({
